eventapp/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==3.0.4
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==3.0.2
distlib==0.3.8
Django==4.2
djangorestframework==3.15.2
filelock==3.16.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
platformdirs==4.3.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
service-identity==24.1.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.4
zope.interface==7.0.3
